+++ beepersource/config.go
@@ -33,6 +33,7 @@ type MatrixConfig struct {
 	PrefixSender            bool
 	PlatformAvatars         bool
 	DMParticipantAvatars    bool
+	AvatarBadges            bool
 	ForceAvatarSync         bool
 	Spaces                  bool
 	InsecureSkipTLS         bool
@@ -74,6 +75,7 @@ func DefaultConfig() Config {
 			PrefixSender:            envBool("BEEPER_MATRIX_PROXY_MATRIX_PREFIX_SENDER", true),
 			PlatformAvatars:         envBool("BEEPER_MATRIX_PROXY_MATRIX_PLATFORM_AVATARS", false),
 			DMParticipantAvatars:    envBool("BEEPER_MATRIX_PROXY_MATRIX_DM_PARTICIPANT_AVATARS", true),
+			AvatarBadges:            envBool("BEEPER_MATRIX_PROXY_MATRIX_AVATAR_BADGES", true),
 			ForceAvatarSync:         envBool("BEEPER_MATRIX_PROXY_MATRIX_FORCE_AVATAR_SYNC", false),
 			Spaces:                  envBool("BEEPER_MATRIX_PROXY_MATRIX_SPACES", false),
 			InsecureSkipTLS:         envBool("BEEPER_MATRIX_PROXY_MATRIX_INSECURE_TLS", false),